Se refactoriza nombres, mueve de ubicacion ficheros de userReducer y actions, actions&Thunks creado
