+++ src/redux/reducers/user/actions&Thunks.js
@@ -1,0 +1,6 @@
+export const authUser = createAsyncThunk(
+    'user/auth',
+    async (body) => {
+        const data = await loginApiCall(body)
+        return data
+    })
