+++ src/redux/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit'
-import counterSlice from './reducers/userSlice'
+import userSlice from './reducers/user/userSlice'
 
 export const STORE = configureStore({
   reducer: {
-    user: counterSlice.reducer
+    user: userSlice.reducer
   }
 })
